ci(htmlproofer): Remove --assume-extension altogether
It seems that it's already set by default.

Also update the description of the CI job to make it clear that the primary purpose of running htmlproofer in the first place is to check for broken internal links. We could check external links too, but it opens a whole other can of worms (as already seen from previous commits), such as 429 (too many request) errors from GitHub and StackOverflow, 404 errors from fonts.gstatic.com, and so on